SCRUM-82 Repository and service class added in Program.cs.
