+++ EntityLayer/Entities/Zone.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EntityLayer.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
