+++ EntityLayer/Interfaces/IAssetPositionHistory.cs
@@ -4,16 +4,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace EntityLayer.Entities
+namespace EntityLayer.Interfaces
 {
-    public interface IAsset
+    public interface IAssetPositionHistory
     {
         public int Id { get; set; }
-        public string Name { get; set; }
+        public int AssetId { get; set; }
         public double X { get; set; }
         public double Y { get; set; }
-        public DateTime LastSync { get; set; }
+        public DateTime DateTime { get; set; }
         public int FloorMapId { get; set; }
-        public bool Active { get; set; }
     }
 }
