+++ EntityLayer/Interfaces/IZone.cs
@@ -1,11 +1,12 @@
-﻿using System;
+﻿using EntityLayer.Entities;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Text.Json.Nodes;
 using System.Threading.Tasks;
 
-namespace EntityLayer.Entities
+namespace EntityLayer.Interfaces
 {
     public interface IZone
     {
