+++ api/Program.cs
@@ -18,10 +18,11 @@ builder.Services.AddSwaggerGen();
 builder.Services.AddSingleton<IAssetRepository, MockAssetRepository>();
 builder.Services.AddSingleton<IZoneRepository, MockZoneRepository>();
 builder.Services.AddSingleton<IFloorMapRepository, MockFloorMapRepository>();
+builder.Services.AddSingleton<IAssetPositionHistoryRepository, MockAssetPositionHistoryRepository>();
 builder.Services.AddSingleton<IAssetService, AssetService>();
 builder.Services.AddSingleton<IZoneService, ZoneService>();
 builder.Services.AddSingleton<IFloorMapService, FloorMapService>();
-
+builder.Services.AddSingleton<IAssetPositionHistoryService, AssetPositionHistoryService>();
 
 var app = builder.Build();
